Move series sharding code to mimirpb, to avoid import cycles.
