+++ pkg/ingester/owned_series.go
@@ -12,7 +12,7 @@ import (
 	"github.com/grafana/dskit/services"
 	"github.com/prometheus/prometheus/model/labels"
 
-	"github.com/grafana/mimir/pkg/util"
+	"github.com/grafana/mimir/pkg/mimirpb"
 )
 
 // Distributor uses WriteNoExtend, but we include all instance states in our operation. Reason is that we want to detect changes to the ring
@@ -169,6 +169,6 @@ func (oss *ownedSeriesService) ownedSeriesUpdateForTenant(userID string, db *use
 
 func SecondaryTSDBHashFunctionForUser(userID string) func(labels.Labels) uint32 {
 	return func(ls labels.Labels) uint32 {
-		return util.ShardByAllLabels(userID, ls)
+		return mimirpb.ShardByAllLabels(userID, ls)
 	}
 }
