+++ pkg/storage/tsdb/util.go
@@ -8,15 +8,15 @@ package tsdb
 import (
 	"github.com/oklog/ulid"
 
-	"github.com/grafana/mimir/pkg/util"
+	"github.com/grafana/mimir/pkg/mimirpb"
 )
 
 // HashBlockID returns a 32-bit hash of the block ID useful for
 // ring-based sharding.
 func HashBlockID(id ulid.ULID) uint32 {
-	h := util.HashNew32()
+	h := mimirpb.HashNew32()
 	for _, b := range id {
-		h = util.HashAddByte32(h, b)
+		h = mimirpb.HashAddByte32(h, b)
 	}
 	return h
 }
